Rename app_ux to ux and move helper functions
