+++ docker/server/index.js
@@ -41,7 +41,7 @@ app.post("/config", (req, res) => {
     api_key,
 
     // App
-    app_ux,
+    ux,
     app_title,
     enable_app_header,
     enable_app_footer,
@@ -93,7 +93,7 @@ app.post("/config", (req, res) => {
     api_key,
 
     // App
-    app_ux,
+    ux,
     app_title,
     enable_app_header,
     enable_app_footer,
@@ -139,5 +139,6 @@ app.post("/config", (req, res) => {
 });
 
 app.listen(port, () => {
+  xw
   console.log(`Example app listening at http://localhost:${port}`);
 });
