+++ src/views/search/SearchView.tsx
@@ -18,41 +18,7 @@ import { useConfigContext } from "../../contexts/ConfigurationContext";
 import { HistoryDrawer } from "./controls/HistoryDrawer";
 import "./searchView.scss";
 import { DeserializedSearchResult } from "./types";
-
-const reorderSummaryCitations = (unorderedSummary: string) => {
-  const allCitations = unorderedSummary.match(/\[\d+\]/g) || [];
-
-  const uniqueCitations = [...new Set(allCitations)];
-  const citationToReplacement: { [key: string]: string } = {};
-  uniqueCitations.forEach((citation, index) => {
-    citationToReplacement[citation] = `[${index + 1}]`;
-  });
-
-  return unorderedSummary.replace(
-    /\[\d+\]/g,
-    (match) => citationToReplacement[match]
-  );
-};
-
-const reorderSearchResults = (
-  searchResults: DeserializedSearchResult[],
-  unorderedSummary: string
-) => {
-  const newSearchResults: DeserializedSearchResult[] = [];
-  const allCitations = unorderedSummary.match(/\[\d+\]/g) || [];
-
-  const addedIndices = new Set<number>();
-  for (let i = 0; i < allCitations.length; i++) {
-    const citation = allCitations[i];
-    const index = Number(citation.slice(1, citation.length - 1)) - 1;
-
-    if (addedIndices.has(index)) continue;
-    newSearchResults.push(searchResults[index]);
-    addedIndices.add(index);
-  }
-
-  return newSearchResults;
-};
+import { reorderSearchResults, reorderSummaryCitations } from "./reorderCitations";
 
 export const SearchView = () => {
   const { isConfigLoaded, app } = useConfigContext();
